refactor: :recycle: GuideInfo 구조 변경
