+++ src/pages/GuideInfo/index.tsx
@@ -1,0 +1,12 @@
+import GuideInfoTab from '@/pages/GuideInfo/components/GuideInfoTab';
+
+const GuideInfo = () => {
+  return (
+    <div className="px-6 py-10 text-center">
+      <div className="mb-4 text-xl text-center">예식정보 및 안내사항</div>
+      <GuideInfoTab />
+    </div>
+  );
+};
+
+export default GuideInfo;
